Added project and mod folder creation
Added the first actual feature of the app, the ability to create a project folder.
Also added a mods folder because I'm planning on having mod support for the app.
Also added a config file that holds some values for the app.
I might change the file format of the app so that it won't be as easy to mess with.